requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
DrissionPage==4.0.5.6
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
redis==5.0.1
gunicorn==21.2.0

